rename cloudcafe package to avoid conflict
