+++ autoscale_cloudcafe/autoscale_fixtures/models/servers.py
