+++ autoscale_cloudroast/test_repo/autoscale/auth/test_otter_rbac_roles.py
@@ -8,7 +8,7 @@ from cloudcafe.identity.v2_0.tokens_api.behaviors import \
 from cloudcafe.identity.v2_0.tokens_api.client import \
     TokenAPI_Client as OSTokenAPI_Client
 from cloudcafe.auth.config import UserConfig
-from autoscale.client import AutoscalingAPIClient
+from autoscale_fixtures.client import AutoscalingAPIClient
 
 
 class OtterRbacTests(AutoscaleFixture):
