+++ autoscale_cloudroast/test_repo/autoscale/functional/launch_config/test_launch_config_negative.py
@@ -3,7 +3,7 @@ Test to negative cases for launch config.
 """
 import unittest
 
-from autoscale.models.request.autoscale_requests import null
+from autoscale_fixtures.models.request.autoscale_requests import null
 from test_repo.autoscale.fixtures import AutoscaleFixture
 from cloudcafe.common.tools.datagen import rand_name
