+++ autoscale_cloudroast/test_repo/autoscale/functional/scaling_group/test_scaling_group_negative.py
@@ -2,7 +2,7 @@
 Test negative scenarios for a scaling group.
 """
 from test_repo.autoscale.fixtures import AutoscaleFixture
-from autoscale.status_codes import HttpStatusCodes
+from autoscale_fixtures.status_codes import HttpStatusCodes
 
 
 class ScalingGroupNegative(AutoscaleFixture):
